[TASK] Move message-related functions to `MessageService`
