+++ Classes/AssetHandler/Html/DataAttributesAssetHandler.php
@@ -17,6 +17,7 @@ use Romm\Formz\AssetHandler\AbstractAssetHandler;
 use Romm\Formz\Core\Core;
 use Romm\Formz\Error\FormResult;
 use Romm\Formz\Form\FormInterface;
+use Romm\Formz\Service\MessageService;
 use TYPO3\CMS\Extbase\Error\Result;
 use TYPO3\CMS\Extbase\Reflection\ObjectAccess;
 
@@ -116,8 +117,8 @@ class DataAttributesAssetHandler extends AbstractAssetHandler
                 $result[self::getFieldDataErrorKey($fieldName)] = '1';
 
                 foreach ($fieldResult->getErrors() as $error) {
-                    $validationName = Core::get()->getMessageValidationName($error);
-                    $messageKey = Core::get()->getMessageKey($error);
+                    $validationName = MessageService::get()->getMessageValidationName($error);
+                    $messageKey = MessageService::get()->getMessageKey($error);
 
                     $result[self::getFieldDataValidationErrorKey($fieldName, $validationName, $messageKey)] = '1';
                 }
