+++ Classes/AssetHandler/JavaScript/FormRequestDataJavaScriptAssetHandler.php
@@ -14,6 +14,7 @@
 namespace Romm\Formz\AssetHandler\JavaScript;
 
 use Romm\Formz\Core\Core;
+use Romm\Formz\Service\MessageService;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Error\Error;
 use TYPO3\CMS\Extbase\Error\Result;
@@ -108,8 +109,8 @@ JS;
 
                     foreach ($formFieldsResult[$fieldName]->getErrors() as $error) {
                         /** @var Error $error */
-                        $validationName = Core::get()->getMessageValidationName($error);
-                        $messageKey = Core::get()->getMessageKey($error);
+                        $validationName = MessageService::get()->getMessageValidationName($error);
+                        $messageKey = MessageService::get()->getMessageKey($error);
 
                         $fieldsErrors[$fieldName][$validationName] = [$messageKey => $error->render()];
                     }
