+++ Classes/Core/Core.php
@@ -13,7 +13,6 @@
 
 namespace Romm\Formz\Core;
 
-use Romm\Formz\Error\FormzMessageInterface;
 use Romm\Formz\Service\TypoScriptService;
 use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
 use TYPO3\CMS\Core\SingletonInterface;
@@ -21,7 +20,6 @@ use TYPO3\CMS\Core\Utility\ArrayUtility;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\PathUtility;
-use TYPO3\CMS\Extbase\Error\Message;
 use TYPO3\CMS\Extbase\Object\ObjectManager;
 use TYPO3\CMS\Extbase\Object\ObjectManagerInterface;
 use TYPO3\CMS\Extbase\Service\EnvironmentService;
@@ -274,36 +272,6 @@ class Core implements SingletonInterface
         }
 
         return $string;
-    }
-
-    /**
-     * Returns the validation name of a message: if it is an instance of
-     * `FormzMessageInterface`, we can fetch it, otherwise `unknown` is
-     * returned.
-     *
-     * @param Message $message
-     * @return string
-     */
-    public function getMessageValidationName(Message $message)
-    {
-        return ($message instanceof FormzMessageInterface)
-            ? $message->getValidationName()
-            : 'unknown';
-    }
-
-    /**
-     * Returns the key of a message: if it is an instance of
-     * `FormzMessageInterface`, we can fetch it, otherwise `unknown` is
-     * returned.
-     *
-     * @param Message $message
-     * @return string
-     */
-    public function getMessageKey(Message $message)
-    {
-        return ($message instanceof FormzMessageInterface)
-            ? $message->getMessageKey()
-            : 'unknown';
     }
 
     /**
